remove unused ability to include internal members in formatter output
