+++ src/Microsoft.DotNet.Interactive.ApiCompatibility.Tests/ApiCompatibilityTests.Formatting_api_is_not_changed.approved.txt
@@ -158,9 +158,6 @@ Microsoft.DotNet.Interactive.Formatting
     public static ITypeFormatter GetPreferredFormatterFor<T>()
     public static System.Void WriteStartProperty(FormatContext context)
   public class PlainTextFormatter<T> : TypeFormatter<T>, ITypeFormatter<T>, ITypeFormatter
-    public static PlainTextFormatter<T> CreateForAnyEnumerable(System.Boolean includeInternals = False)
-    public static PlainTextFormatter<T> CreateForAnyObject(System.Boolean includeInternals = False)
-    public static PlainTextFormatter<T> CreateForMembers(Expression<Func<T,System.Object>> members)
     .ctor(FormatDelegate<T> format)
     .ctor(Action<T,FormatContext> format)
     .ctor(Func<T,System.String> format)
